requirements.txt
arxiv>=2.1.0
openai>=1.0.0
pyyaml>=6.0
schedule>=1.2.0
pytz>=2024.1